requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0 # Added for production WSGI server
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
# openai==1.3.5 # Removed - Migrating to Vertex AI
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0

# Azure Text-to-Speech (Removed - Migrating to Google Cloud TTS)
# azure-cognitiveservices-speech==1.32.1

# Spotify API
spotipy==2.23.0

# Google AI (Vertex AI)
google-cloud-aiplatform==1.36.4
vertexai
google-cloud-language==2.13.0 # Added for Natural Language API
google-maps-services-python==0.1.0 # Added for Maps APIs
google-cloud-texttospeech==2.14.1 # Added for Google Cloud TTS
google-cloud-storage==2.14.0 # Added for GCS
google-cloud-speech==2.20.0 # Added for Speech-to-Text
google-cloud-secret-manager==2.16.2 # Added for Secret Manager
google-cloud-logging==3.5.0 # Added for Cloud Logging integration
# Caching
redis==5.0.1

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
ruff==0.1.6

# Additional AI and ML dependencies
langchain==0.1.0
langchain-google-vertexai
google-cloud-translate==3.11.2
prometheus-client==0.19.0
alembic==1.13.1

# Third-party integrations
stripe==7.5.0  # For payment processing
twilio==8.10.1  # For SMS notifications
sendgrid==6.11.0  # For email notifications

# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2  # For ML-based personalization

# Async support
aiofiles==23.2.1
aiohttp==3.9.1

# Security enhancements
cryptography==41.0.7
python-owasp==2.0.0

# Monitoring and observability
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Rate limiting and caching
slowapi==0.1.9
python-limiter==1.2.1
